+++ code/version 3/HARDWARE/ADC/adc.h
@@ -15,7 +15,7 @@
 
 void Adc_Init(void);
 u16  Get_Adc(u8 ch); 
-u16 Get_Adc_Average(u8 ch,u8 times);
+u16 Get_Adc_Average(u16 ch,u16 times);
 extern float temp_avrg;
 extern int shidu;
 extern int soil_moisture;
